back/requirements.txt
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
backports.zoneinfo==0.2.1
django-filter
psycopg2
psycopg2-binary
Pillow
wheel
sqlparse
setuptools
PyJWT
pytz
django-cors-headers